refactor: 코드 리팩토링 및 정리
