+++ src/App.js
@@ -1,6 +1,6 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
 import { GAME_MSG } from "./Messages";
-import { compareNum } from "./BaseballGame";
+import { compareNum } from "./baseballGame/baseballGame";
 
 class App {
   async play() {
